% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{package}
\name{ProjectDataViz}
\alias{ProjectDataViz}
\alias{portaldata-package}
\title{Portal Data Set and Analysis}
\format{
A data frame with 4 columns and 1000 rows, The names of the columns are: "Year","Class","Metric", and "Number.of.Species"
"Single_Measure"
#' Portal Data Set and Analysis

This package contains data sets from the ProjectDataViz Project
for easy reuse and reanalysis.
}
\source{

}
\description{
This package contains data sets from the ProjectDataViz Project
for easy reuse and reanalysis.

A data set containing the taxonomy of species sampled for the Portal Project Data.
We followed the taxonomy from the Open Tree of Life .
}
